Add rate to comment, fix schema

+++ app/validators/api/v1/meal_schema.rb
@@ -13,7 +13,7 @@ module Api
         required(:time, :integer).filled
         required(:servings, :integer).filled
         required(:products).schema do
-          required(:id, :bigint).filled
+          required(:id, :integer).filled
           required(:amount, :integer).filled
         end
       end
